pyproject.toml
[project]
name = ""
version = "0.0.1"
requires-python = "~=3.12"
dependencies = [
    "azure-core",
    "azure-cosmos~=4.3.1",
    "azure-functions",
    "azure-monitor-opentelemetry",
    "cachecontrol",
    "cloudscraper>=1.2.71",
    "coverage",
    "curl-cffi>=0.14.0",
    "google-auth-oauthlib",
    "jinja2==3.1.6",
    "pydantic[email]",
    "six>=1.17.0",
]

[tool.black]
line-length = 90
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[dependency-groups]
dev = ["black"]

[tool.uv]
package = false

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
